Moving the client library to packages explicitly.
